Refactor: Move filtering to dedicated file
